+++ test/jest/fixtures/instance.js
@@ -1,0 +1,22 @@
+export const instance = {
+  id: 'cf23adf0-61ba-4887-bf82-956c4aae2260',
+  title: 'ADVANCING LIBRARY EDUCATION: TECHNOLOGICAL INNOVATION AND INSTRUCTIONAL DESIGN',
+  instanceTypeId: 'abc5dd49-7ede-49cb-84db-6fac770279a8',
+  identifiers: [
+    {
+      identifierTypeId: '8261054f-be78-422d-bd51-4ed9f33c3422',
+      value: '9781466636897',
+    },
+  ],
+  contributors: [
+    {
+      'contributorNameTypeId': '2b94c631-fca9-a892-c730-03ee529ffe2a',
+      'name': 'Samuels, Simon',
+    },
+  ],
+  editions: ['edition'],
+  publication: [{
+    publisher: 'Samuels, Simon',
+    dateOfPublication: '2021',
+  }],
+};